many: rename styles as <component>.styles.js
