+++ src/components/auth/RegisterModalsContainer/RegisterModalsContainer.styles.js
