+++ src/components/auth/VideoAgreeModal/VideoAgreeModal.styles.js
