+++ src/components/auth/VideoModal/VideoModal.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import styles from './VideoModal-style.js';
+import styles from './VideoModal.styles';
 
 import { Text, View } from 'react-native';
