+++ src/components/auth/VideoModal/VideoModal.styles.js
